bg(redundancy): remove redendant data
- Use same helpers, middleware for both version 1 & 2
- Use 204 HTTP status code for deletion rather than 200
- Fix errors in documention
- Fix get all entries response when there is no entry found

[Finishes #169572238]
