--- Server/V1/Model/mockData/testEntry.js
@@ -1,19 +1,0 @@
-const testEntry = [
-    {
-        title: 'My title',
-        description: 'My description',
-    }, {
-        title: true,
-        description: 'hey there',
-    }, {
-        title: 'Le 21 dec 2019',
-        description: 'my new diary',
-    }, {
-        entryId: 1,
-    }, {
-        entryId: -1,
-    }, {
-        entryId: 500,
-    },
-];
-export default testEntry;
