--- Server/V1/Model/mockData/testUSer.js
@@ -1,30 +1,0 @@
-const testUsers = [{
-    firstName: 'testUser',
-    lastName: 'testUser',
-    email: 'testUser@gmail.com',
-    password: 'password',
-}, {
-    lastName: 'lastName',
-    firstName: '',
-    email: 'userEmail@gmail.com',
-    password: 'password',
-}, {
-    email: 'testUser@gmail.com',
-    password: 'password',
-}, {
-    email: 'testUser1@gmail.com',
-    password: 'password',
-}, {
-    email: 'testUser@gmail.com',
-    password: 'password1',
-}, {
-    email: 'testUser@gmail.com',
-    // password: 'password',
-}, {
-    firstName: 'testUser2',
-    lastName: 'testUser2',
-    email: 'testUser2@gmail.com',
-    password: 'password',
-},
-];
-export default testUsers;
