+++ Server/V2/Test/v_entryTests.test.js
@@ -3,7 +3,7 @@ import chaiHttp from 'chai-http';
 import { describe, it } from 'mocha';
 import dotenv from 'dotenv';
 import app from '../../../index';
-import testEntry from '../Model/mockData/testEntry';
+import testEntry from '../../mockData/testEntry';
 
 dotenv.config();
 const { expect } = chai;
@@ -122,19 +122,6 @@ describe('Modify an entry', () => {
             });
     });
 });
-describe('Get all entries', () => {
-    it('Should should return all entries', (done) => {
-        chai.request(app).get('/api/v2/entries')
-            .set('Authorization', process.env.userToken1)
-            .end((err, res) => {
-                expect(res).to.have.status(200);
-                expect(res.body).to.have.property('data');
-                expect(res.body.data).to.be.a('object');
-                expect(res.body.message).to.equal('Entries successfully retreived');
-                done();
-            });
-    });
-});
 describe('Get specific entry', () => {
     it('Should not return an entry : entryId not found', (done) => {
         chai.request(app).get(`/api/v2/entries/${testEntry[5].entryId}`)
@@ -173,8 +160,30 @@ describe('Delete an entry', () => {
         chai.request(app).delete(`/api/v2/entries/${testEntry[3].entryId}`)
             .set('Authorization', process.env.userToken1)
             .end((err, res) => {
+                expect(res).to.have.status(204);
+                done();
+            });
+    });
+});
+describe('Get all entries', () => {
+    it('Should should return all entries', (done) => {
+        chai.request(app).get('/api/v2/entries')
+            .set('Authorization', process.env.userToken2)
+            .end((err, res) => {
                 expect(res).to.have.status(200);
-                expect(res.body.message).to.equal('Entry successfully deleted');
+                expect(res.body).to.have.property('data');
+                expect(res.body.data).to.be.a('object');
+                expect(res.body.message).to.equal('Entries successfully retreived');
+                done();
+            });
+    });
+    it('Should should not return an entry: no entry found', (done) => {
+        chai.request(app).get('/api/v2/entries')
+            .set('Authorization', process.env.userToken1)
+            .end((err, res) => {
+                expect(res).to.have.status(404);
+                expect(res.body).to.have.property('error');
+                expect(res.body.error).to.equal('No entry found');
                 done();
             });
     });
