--- Server/V2/helpers/checkOwner.js
@@ -1,7 +1,0 @@
-const checkOwner = (req, entryId) => {
-    if (entryId !== req.tokenData.id) {
-        return false;
-    }
-    return true;
-};
-export default checkOwner;
