--- Server/V2/helpers/checkPassword.js
@@ -1,4 +1,0 @@
-import bcrypt from 'bcrypt';
-
-const checkPassword = (password, hashedPassword) => bcrypt.compareSync(password, hashedPassword);
-export default checkPassword;
