--- Server/V2/helpers/generateToken.js
@@ -1,8 +1,0 @@
-import jwt from 'jsonwebtoken';
-
-const generateToken = (id, email) => jwt.sign({
-    id, email,
-}, process.env.secret, {
-    expiresIn: '7d',
-});
-export default generateToken;
