+++ Server/helpers/checkPassword.js
