+++ Server/helpers/generateToken.js
