+++ Server/helpers/hashPassword.js
